# System.Linq Orderby performance optimization