@stephentoub 2021-01-04
Do you have numbers for the case where there is a secondary sort? e.g. a small array with a trivial primary and secondary sort?  In that case we're paying for the virtual dispatch and reflection checks but not getting anything for them.