@johnthcall 2021-01-04
@stephentoub
For 10 strings containing random int's 
StringToList is Strings.OrderBy(s => s).ThenBy(s => s).ToList();
StringToArray is Strings.OrderBy(s => s).ThenBy(s => s).ToArray();
|        Method | Toolchain |  N |     Mean |     Error | Ratio | Allocated |
|-------------- |---------- |--- |---------:|----------:|------:|----------:|
|  StringToList |       new | 10 | 2.880 us | 0.2544 us |  1.00 |     864 B |
|  StringToList |  upstream | 10 | 2.801 us | 0.2216 us |  0.99 |     864 B |
|               |           |    |          |           |       |           |
| StringToArray |       new | 10 | 2.913 us | 0.2185 us |  1.00 |     832 B |
| StringToArray |  upstream | 10 | 2.934 us | 0.2236 us |  1.01 |     832 B |