@stephentoub 2021-03-16
> If you don't have any further suggestion I'll abandon this PR.

Thanks for all your efforts to try to improve things here, @johnthcall.  Given the data, let's go ahead and close this.